+++ src/main/java/com/workhub/userTable/api/UserAdminApi.java
@@ -4,16 +4,21 @@ import com.workhub.global.response.ApiResponse;
 import com.workhub.userTable.dto.user.request.AdminPasswordResetRequest;
 import com.workhub.userTable.dto.user.request.UserRegisterRecord;
 import com.workhub.userTable.dto.user.request.UserRoleUpdateRequest;
+import com.workhub.userTable.dto.user.response.UserDetailResponse;
+import com.workhub.userTable.dto.user.response.UserListResponse;
 import com.workhub.userTable.dto.user.response.UserTableResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+
+import java.util.List;
 
 @Tag(name = "관리자 사용자 관리", description = "관리자 전용 사용자 생성/권한 관리/삭제 API")
 public interface UserAdminApi {
@@ -94,4 +99,66 @@ public interface UserAdminApi {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "500", description = "서버 오류 (회원 삭제 실패)")
     })
     ResponseEntity<ApiResponse<Object>> deleteUser(Long userId);
+
+    @Operation(
+            summary = "회원 목록 조회",
+            description = "모든 회원 정보를 조회합니다."
+    )
+    @ApiResponses({
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(
+                    responseCode = "200",
+                    description = "회원 목록 조회 성공",
+                    content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            array = @ArraySchema(schema = @Schema(implementation = UserListResponse.class)))
+            ),
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(
+                    responseCode = "401",
+                    description = "인증 실패",
+                    content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            schema = @Schema(implementation = ApiResponse.class))
+            ),
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(
+                    responseCode = "500",
+                    description = "서버 오류 (회원 목록 조회 실패)",
+                    content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            schema = @Schema(implementation = ApiResponse.class))
+            )
+    })
+    List<UserListResponse> getUserTable();
+
+    @Operation(
+            summary = "회원 상세 조회",
+            description = "특정 회원의 상세 정보를 조회합니다.",
+            parameters = {
+                    @Parameter(name = "userId", description = "상세 정보를 확인할 회원 식별자", in = ParameterIn.PATH, required = true)
+            }
+    )
+    @ApiResponses({
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(
+                    responseCode = "200",
+                    description = "회원 상세 조회 성공",
+                    content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            schema = @Schema(implementation = UserDetailResponse.class))
+            ),
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(
+                    responseCode = "401",
+                    description = "인증 실패",
+                    content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            schema = @Schema(implementation = ApiResponse.class))
+            ),
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(
+                    responseCode = "404",
+                    description = "사용자를 찾을 수 없음",
+                    content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            schema = @Schema(implementation = ApiResponse.class))
+            ),
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(
+                    responseCode = "500",
+                    description = "서버 오류 (회원 상세 조회 실패)",
+                    content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            schema = @Schema(implementation = ApiResponse.class))
+            )
+    })
+    ResponseEntity<ApiResponse<UserDetailResponse>> getUser(Long userId);
+
 }
